Promote `context` module to separate dir
